requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
faiss-cpu
pypdf